+++ packet/ImagePacket.h
@@ -14,7 +14,7 @@ public:
 	void incSeqNumber();
 	int16_t getSeqNumber();
 
-	void setImageDataLength(uint8_t imageDataLength);
+	void setImageDataLength(uint16_t imageDataLength);
 	void setImageDataByte(uint8_t value, int16_t bytePos);
 	uint8_t getImageDataByte(int16_t bytePos);
 	uint16_t toBinary(uint8_t* imagePacketBinaryBuffer, int16_t bufferSize);
@@ -22,7 +22,7 @@ public:
 private:
 	uint16_t _teamId;
 	int16_t _seqNumber;
-	uint8_t _imageDataLength;
+	uint16_t _imageDataLength;
 	uint8_t _imageData[IMAGE_PACKET_DATA_SIZE];
 	
 };
